requirements.txt
# Core Framework
rasa==3.6.11
rasa-sdk==3.6.2

# LLM & Transformers
torch==2.1.0
torchvision==0.16.0
torchaudio==0.16.0
transformers==4.34.1
accelerate==0.24.1
bitsandbytes==0.41.2
peft==0.7.1
xformers==0.0.22

# LLM Runner
ollama==0.0.39
langchain==0.1.1
langchain-community==0.0.5

# Data Processing & Utils
python-dotenv==1.0.0
pydantic==2.5.0
PyYAML==6.0.1
requests==2.31.0
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# API & Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
werkzeug==3.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
typing-extensions==4.8.0
